requirements.txt
pickleshare==0.7.5
streamlit==0.61.0
Pillow==7.1.2